Update WideRisingStyle.user.css
Lots of stuff going on:

-Reorganized the document to make it easier to navigate. Makes working on it easier, and lets the user see the code more clearly in case they want to tweak something manually beyond the options already given by the userstyle.

-Fixed the Forums page, now it should adjust according to the width chosen by the user just like the rest of the site.

-Added the affectionately called "Silly Mode": a small collection of additional features made just for fun. As of now, Silly Mode does the following: 1) Replaces the adblock warning image ("Flight Rising relies on adverts to run, etc etc...") with pure CSS text for the sake of aesthetics, since the original image is very low-res and hard to read. 2) Hides a random separator image in the Nesting Grounds that's ugly as hell and I don't even know why it was there in the first place ??. 3) Adds a top margin to the new hatchling screen in the Nesting Grounds to let the UI breathe a little, especially when widening the layout.
There's an extra feature that was originally meant for Silly Mode but I decided to make a separate style for in case people find it helpful and don't want any of the other stuff. It shows what materials get turned into what at Baldwin's Bubbling Brew, and you can download it here: https://github.com/JustDownloadin/BaldwinMatsHelper  . I may include it in Wide Rising in the future though, we'll see. Most of the code is already done, it'd just be a matter of how to integrate the options themselves to this style (if they should be any different at all to begin with).